+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Support `CreateSplitButton` extension.
 - Support `CreatePulldownButton` extension.
+- Support `CreatePanel` extension.
 ### UI
 - Update `RibbonExtension` to support `CreateSplitButton` and `CreatePulldownButton`.
 - Update `MessageBox.ShowMessage` to close window using `Key.Escape`.
 - Add `SetListImageSize` to set the image size for `RibbonListButton`.
+- Update `CreatePanel` and `CreateOrSelectPanel` to use `TabName` and `PanelName` parameters.
 ### Example
 - Add `CreateSplitButton` and `CreatePulldownButton` to the example project.